Add unicode representation, along with smart switching
There's both the traditional shortcode (:black_large_square:) and the now-added unicode () representation.
The program decided which one to send to the Discord servers by looking at which representation is able to encode
more characters per line: Discord has a limit of 2000 characters for shortcode representation, and 198 for emoji
representation. Sometimes the shortcode representation will be above its limit and the emoji representation will go
through, and other times it's vice versa. This hybrid approach ensures that we're always able to send the longest
possible message.